Extract autoCompletePath from lazy
